requirements.txt
asgiref==3.3.4
autopep8==1.5.7
dj-database-url==0.5.0
Django==3.2.3
django-heroku==0.0.0
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
whitenoise==5.2.0
